common----修改项目必读,修改字段类型,添加枚举类型

+++ mall-common/src/main/java/com/mall/common/request/ItemRequest.java
@@ -51,7 +51,7 @@ public class ItemRequest extends Base {
     /**
      * 所属类目，叶子类目
      */
-    private Integer itemCatId;
+    private Long itemCatId;
 
     /**
      * 商品描述
@@ -119,11 +119,11 @@ public class ItemRequest extends Base {
         this.image = image;
     }
 
-    public Integer getItemCatId() {
+    public Long getItemCatId() {
         return itemCatId;
     }
 
-    public void setItemCatId(Integer itemCatId) {
+    public void setItemCatId(Long itemCatId) {
         this.itemCatId = itemCatId;
     }
